requirements.txt
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
requests>=2.31.0
ollama
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc